TP-80 Moves Routes config to Common directory. Fixes routes.
