+++ src/App.js
@@ -17,7 +17,7 @@ import { InMemoryCache } from 'apollo-cache-inmemory';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import { faSearch, faBars } from '@fortawesome/free-solid-svg-icons';
 import { faTwitter, faInstagram } from '@fortawesome/free-brands-svg-icons';
-import Routes from './RoutesConfig';
+import Routes from './components/Common/RoutesConfig';
 
 const httpLink = createHttpLink({
   uri: 'http://back.3blmedia.com/graphql'
@@ -50,16 +50,12 @@ class App extends Component {
               <ApolloProvider client={client}>
                 <Header site_title={this.state.pageTitle}/>
                 <Route exact path={'/'} render={(props) => <Home {...props} newTitle={this.setPageTitle} />}/>
-                <Route path={Routes.STORIES} component={StoriesPage}/>
-                <Route path={Routes.FEED} render={(props) => <TheFeedPage {...props} />}/>
-                <Route path={Routes.EDIOTRS} component={HomeEditors}/>
+                <Route path={Routes.THEFEED} render={(props) => <TheFeedPage {...props} />}/>
+                <Route path={Routes.EDITORS} component={HomeEditors}/>
                 <Route exact path={Routes.SPONSOREDSERIES} component={SponsoredSeriesAll}/>
-                <Route path={Routes.PRIMESTORIES} component={""}/>
-                <Route path={Routes.SINGLESERIES} component={""}/>
-                <Route path={Routes.STORYALL} component={""}/>
-                <Route path={Routes.STORYDETAIL} component={""}/>
-                <Route path={Routes.CONTACT} component={""}/>
-                <Route path={Routes.BTS} component={""}/>
+                <Route path={Routes.SINGLESPONSOREDSERIES} component={""}/>
+                <Route path={Routes.STORIES} component={StoriesPage}/>
+                <Route path={Routes.SINGLESTORY} component={""}/>
                 <Footer/>
               </ApolloProvider>
             </div>
