--- src/RoutesConfig.js
@@ -1,16 +1,0 @@
-const Routes = {
-    "PRIMESTORIES":"prime-stories",
-    "STORYALL":'/storyall',
-    "STORIES":"/stories",
-    "SPONSOREDSERIES": "/sponsored-series",
-    "SINGLESERIES":"/single-series",
-    "EDIOTRS": "/editors",
-    "EDITORDETAILS":"/editordatails/:id",
-    "FEED":"/the-feed",
-    "BTS":"/bts",
-    "CONTACT":"/contact",
-    "STORYDETAIL":"/storydetail/:id",
-    "SEARCH":"/search"
-}
-
-export default Routes;
