+++ src/components/Common/RoutesConfig.js
@@ -1,0 +1,13 @@
+const Routes = {
+    "SPONSOREDSERIES": "/sponsored-series",
+    "SINGLESPONSOREDSERIES":"/sponsored-series/:id",
+    "EDITORS": "/editors",
+    "SINGLEEDITOR":"/editor/:id",
+    "THEFEED":"/the-feed",
+    "CONTACT":"/contact",
+    "STORIES":"/stories",
+    "SINGLESTORY":"/story/:id",
+    "SEARCH":"/search"
+};
+
+export default Routes;
